+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ## Component Enhancements
 
 * `Link` component now has a prop of `iconAlign` to align icons to the right of the link's text.
-* `Row` component can now be given a custom value for the size of the gutter using the prop `gutter`.
+* `Row` component can now be given a size to control the size of the gutter using the prop `gutter` (eg. `extra-small`, `small`, `medium`, `large` or `extra-large`).
 
 ## Minor Improvements